Application Dashboard with Shiny(R)
Application for exploratory data analysis and presentation of simple linear regression, as well as providing basic descriptive statistics for the selected dataset. Since EDA can sometimes be time-consuming due to the sheer number of variables in a data set, writing code for each graph or group of graphs. The application also contains functions for displaying the results of the research regression calculation and its visualization. The user will be able to choose the type of graph, the variables that will be displayed, it will also be possible to calculate the basic descriptive statistics for the columns, save the resulting graphs, as well as change the dataset.